修改nrs1800 makefile，增加sroute工具相关读写寄存器选项

+++ files/platform-2.0.0/main/tools/nrs1800/nr_config.h
@@ -9,6 +9,6 @@
 #endif
 
 extern void nr_config_init(void);
-extern srio_route_rule  **nr_get_route_table(int *count);
+extern srio_route_rule  **nr_get_route_table(size_t *count);
 
 #endif
